Migrate to 12.6+ resource
